React Router Dom Implementation and other additions
Firstly, I have finished creating a homepage and begun with the implementation of each project having an empty page that they can work with - a blank canvas. Furthermore, I need to add more of the use navigation functions (particularly in the navbar as well as others). Also, I need to implement a footer.